requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
pyjwt==1.7.1